Refactor rendering pipeline and reorganize packages
