+++ gradle.properties
@@ -69,9 +69,9 @@ usesMixins = true
 # Adds some debug arguments like verbose output and export
 usesMixinDebug = true
 # Specify the location of your implementation of IMixinConfigPlugin. Leave it empty otherwise.
-mixinPlugin = mixins.early.HexTextEarlyMixins
+mixinPlugin = mixin.config.HexTextMixinPlugin
 # Specify the package that contains all of your Mixins. You may only place Mixins in this package or the build will fail!
-mixinsPackage = mixins
+mixinsPackage = mixin
 # Specify the core mod entry class if you use a core mod. This class must implement IFMLLoadingPlugin!
 # This parameter is for legacy compatibility only
 # Example value: coreModClass = asm.FMLPlugin + modGroup = com.myname.mymodid -> com.myname.mymodid.asm.FMLPlugin
