--- src/main/java/kamkeel/hextext/client/TokenHighlight.java
@@ -1,15 +1,0 @@
-package kamkeel.hextext.client;
-
-public class TokenHighlight {
-    public final float x;
-    public final float y;
-    public final float width;
-    public final int color;
-
-    public TokenHighlight(float x, float y, float width, int color) {
-        this.x = x;
-        this.y = y;
-        this.width = width;
-        this.color = color;
-    }
-}
