+++ src/main/java/kamkeel/hextext/client/render/TokenHighlight.java
@@ -1,0 +1,15 @@
+package kamkeel.hextext.client.render;
+
+public class TokenHighlight {
+    public final float x;
+    public final float y;
+    public final float width;
+    public final int color;
+
+    public TokenHighlight(float x, float y, float width, int color) {
+        this.x = x;
+        this.y = y;
+        this.width = width;
+        this.color = color;
+    }
+}
