+++ src/main/java/kamkeel/hextext/client/render/font/FontRendererBridge.java
@@ -1,0 +1,42 @@
+package kamkeel.hextext.client.render.font;
+
+import net.minecraft.client.gui.FontRenderer;
+
+public interface FontRendererBridge {
+
+    FontRenderer getFontRenderer();
+
+    void setRandomStyle(boolean enabled);
+
+    void setBoldStyle(boolean enabled);
+
+    void setStrikethroughStyle(boolean enabled);
+
+    void setUnderlineStyle(boolean enabled);
+
+    void setItalicStyle(boolean enabled);
+
+    void setTextColor(int color);
+
+    int getTextColor();
+
+    float getAlpha();
+
+    float getRedComponent();
+
+    float getBlueComponent();
+
+    float getGreenComponent();
+
+    int[] getColorCodePalette();
+
+    float getPosX();
+
+    float getPosY();
+
+    int getFontHeight();
+
+    void applyColorComponents(float red, float green, float blue, float alpha);
+
+    void resetFormattingStyles();
+}
