+++ src/main/java/kamkeel/hextext/client/render/font/GlyphRenderer.java
@@ -1,0 +1,6 @@
+package kamkeel.hextext.client.render.font;
+
+public interface GlyphRenderer {
+    float renderDefault(int glyphIndex, boolean italic);
+    float renderUnicode(char character, boolean italic);
+}
