+++ src/main/java/kamkeel/hextext/mixin/config/HexTextLateMixinLoader.java
@@ -1,0 +1,29 @@
+package kamkeel.hextext.mixin.config;
+
+import com.gtnewhorizon.gtnhmixins.ILateMixinLoader;
+import com.gtnewhorizon.gtnhmixins.LateMixin;
+import org.spongepowered.asm.mixin.MixinEnvironment;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Set;
+
+@LateMixin
+public final class HexTextLateMixinLoader implements ILateMixinLoader {
+
+    private static final MixinEnvironment.Side SIDE = MixinEnvironment.getCurrentEnvironment().getSide();
+
+    @Override
+    public String getMixinConfig() {
+        return "mixins.hextext.late.json";
+    }
+
+    @Override
+    public List<String> getMixins(Set<String> loadedMods) {
+        List<String> mixins = new ArrayList<>();
+        if (SIDE == MixinEnvironment.Side.CLIENT) {
+            // No late mixins yet
+        }
+        return mixins;
+    }
+}
