--- src/main/java/kamkeel/hextext/mixins/early/impl/client/MixinGuiTextField.java
@@ -1,30 +1,0 @@
-package kamkeel.hextext.mixins.early.impl.client;
-
-import kamkeel.hextext.client.FontRenderContext;
-import net.minecraft.client.gui.GuiTextField;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Unique;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-@Mixin(GuiTextField.class)
-public abstract class MixinGuiTextField {
-
-    @Unique
-    private boolean hextext$legacy$rawPushed;
-
-    @Inject(method = "drawTextBox", at = @At("HEAD"))
-    private void hextext$legacy$beginRawMode(CallbackInfo ci) {
-        FontRenderContext.pushRawTextRendering();
-        hextext$legacy$rawPushed = true;
-    }
-
-    @Inject(method = "drawTextBox", at = @At("RETURN"))
-    private void hextext$legacy$endRawMode(CallbackInfo ci) {
-        if (hextext$legacy$rawPushed) {
-            FontRenderContext.popRawTextRendering();
-            hextext$legacy$rawPushed = false;
-        }
-    }
-}
