--- src/main/java/kamkeel/hextext/mixins/late/HexTextLateMixins.java
@@ -1,29 +1,0 @@
-package kamkeel.hextext.mixins.late;
-
-import com.gtnewhorizon.gtnhmixins.ILateMixinLoader;
-import com.gtnewhorizon.gtnhmixins.LateMixin;
-import org.spongepowered.asm.mixin.MixinEnvironment;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
-
-@LateMixin
-public class HexTextLateMixins implements ILateMixinLoader {
-
-    public static final MixinEnvironment.Side side = MixinEnvironment.getCurrentEnvironment().getSide();
-
-    @Override
-    public String getMixinConfig() {
-        return "mixins.hextext.late.json";
-    }
-
-    @Override
-    public List<String> getMixins(Set<String> loadedMods) {
-        List<String> mixins = new ArrayList<>();
-        if (side == MixinEnvironment.Side.CLIENT) {
-
-        }
-        return mixins;
-    }
-}
